requirements.txt
flask==3.0.2
beautifulsoup4==4.12.3
icalendar==5.0.12
requests==2.31.0
python-dateutil==2.8.2 